+++ dwds/test/integration/events_common.dart
@@ -435,8 +435,13 @@ void testWithDwds({required TestSdkConfigurationProvider provider}) {
         });
 
         tearDown(() async {
-          // Resume execution to not impact other tests.
-          await service.resume(isolateId);
+          // We must resume execution in case a test left the isolate paused, but
+          // error 106 is expected if the isolate is already running.
+          try {
+            await service.resume(isolateId);
+          } on RPCError catch (e) {
+            if (e.code != 106) rethrow;
+          }
         });
 
         test('emits RESUME events', () async {