+++ dwds/CHANGELOG.md
@@ -10,6 +10,7 @@
 - Add sourcemap logic fixes to DDC Library Bundle + build_runner execution scheme.
 - Update pathing logic for Windows and the DDC Library Bundle module system.
 - Fix serialization of `HotRestartRequest` in `AppConnection`.
+- Split integration tests across DDC module systems.
 - Split additional tests across DDC module systems.
 
 ## 27.0.0